Revert "Add To Cart Form: Add `Stepper` layout" (#48624)
* Revert "Add To Cart Form: Improve the default style and add the `Stepper` style (#47664)"

This reverts commit 0712c6037b4e8a868c1e66ff18fc7046e606a23f.

* Add changefile(s) from automation for the following project(s): woocommerce-blocks, woocommerce

---------
