+++ plugins/woocommerce-blocks/assets/js/atomic/blocks/index.js
@@ -13,6 +13,7 @@ import './product-elements/summary';
 import './product-elements/sale-badge';
 import './product-elements/sku';
 import './product-elements/stock-indicator';
+import './product-elements/add-to-cart-form';
 import './product-elements/product-image-gallery';
 import './product-elements/product-details';
 import './product-elements/product-reviews';
