+++ plugins/woocommerce-blocks/assets/js/atomic/blocks/product-elements/add-to-cart-form/editor.scss
@@ -1,0 +1,6 @@
+.wc-block-editor-add-to-cart-form-container {
+	cursor: help;
+	gap: 10px;
+	display: flex;
+	flex-direction: column;
+}
