+++ plugins/woocommerce-blocks/assets/js/atomic/blocks/product-elements/add-to-cart-form/index.tsx
@@ -11,7 +11,6 @@ import metadata from './block.json';
 import edit from './edit';
 import './style.scss';
 import './editor.scss';
-import '../../../base/components/quantity-selector/style.scss';
 
 const blockSettings = {
 	edit,
