+++ plugins/woocommerce-blocks/assets/js/atomic/blocks/product-elements/add-to-cart-form/style.scss
@@ -1,0 +1,33 @@
+.wc-block-add-to-cart-form {
+	width: unset;
+	/**
+	* This is a base style for the input text element in WooCommerce that prevents inputs from appearing too small.
+	*
+	* @link https://github.com/woocommerce/woocommerce/blob/95ca53675f2817753d484583c96ca9ab9f725172/plugins/woocommerce/client/legacy/css/woocommerce-blocktheme.scss#L203-L206
+	*/
+	.input-text {
+		font-size: var(--wp--preset--font-size--small);
+		padding: 0.9rem 1.1rem;
+	}
+
+	.quantity {
+		display: inline-block;
+		float: none;
+		margin-right: 4px;
+		vertical-align: middle;
+
+		.qty {
+			margin-right: 0.5rem;
+			width: 3.631em;
+			text-align: center;
+		}
+	}
+}
+
+.woocommerce div.product .wc-block-add-to-cart-form form.cart .quantity {
+	margin-bottom: 10px;
+}
+
+.woocommerce div.product .wc-block-add-to-cart-form form.cart button.single_add_to_cart_button {
+	margin-bottom: 10px;
+}
