--- plugins/woocommerce-blocks/assets/js/blocks/product-elements/add-to-cart-form/editor.scss
@@ -1,39 +1,0 @@
-.wc-block-editor-add-to-cart-form-container {
-	cursor: help;
-	gap: 10px;
-	display: flex;
-	flex-direction: column;
-	width: 400px;
-}
-
-.wc-block-editor-quantity-selector-style {
-	width: 100%;
-}
-
-.quantity .qty {
-	// WooCommerce core styles: https://github.com/woocommerce/woocommerce/blob/c9f62609155825cd817976c7611b9b0415e90f2f/plugins/woocommerce/client/legacy/css/woocommerce.scss/#L111-L112
-	width: 3.631em;
-	text-align: center;
-}
-
-div.product form.cart div.quantity {
-	display: inline-block;
-	float: none;
-	vertical-align: middle;
-}
-
-input {
-	height: auto;
-}
-
-.wc-block-components-quantity-selector--input {
-	padding: unset;
-}
-
-.wc-block-add-to-cart-form.wc-block-add-to-cart-form--stepper {
-	.wc-block-components-quantity-selector {
-		height: 100%;
-		margin: 0 4px 0 0;
-		background: #fff;
-	}
-}
