--- plugins/woocommerce-blocks/assets/js/blocks/product-elements/add-to-cart-form/settings.tsx
@@ -1,77 +1,0 @@
-/**
- * External dependencies
- */
-import { InspectorControls } from '@wordpress/block-editor';
-import { __ } from '@wordpress/i18n';
-import {
-	PanelBody,
-	// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-	// @ts-ignore - Ignoring because `__experimentalToggleGroupControl` is not yet in the type definitions.
-	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
-	__experimentalToggleGroupControl as ToggleGroupControl,
-	// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-	// @ts-ignore - Ignoring because `__experimentalToggleGroupControl` is not yet in the type definitions.
-	// eslint-disable-next-line @wordpress/no-unsafe-wp-apis
-	__experimentalToggleGroupControlOption as ToggleGroupControlOption,
-} from '@wordpress/components';
-
-export enum QuantitySelectorStyle {
-	Input = 'input',
-	Stepper = 'stepper',
-}
-
-type AddToCartFormSettings = {
-	quantitySelectorStyle: QuantitySelectorStyle;
-	setAttributes: ( attributes: {
-		quantitySelectorStyle: QuantitySelectorStyle;
-	} ) => void;
-};
-
-const getHelpText = ( quantitySelectorStyle: QuantitySelectorStyle ) => {
-	if ( quantitySelectorStyle === QuantitySelectorStyle.Input ) {
-		return __(
-			'Shoppers can enter a number of items to add to cart.',
-			'woocommerce'
-		);
-	}
-	if ( quantitySelectorStyle === QuantitySelectorStyle.Stepper ) {
-		return __(
-			'Shoppers can use buttons to change the number of items to add to cart.',
-			'woocommerce'
-		);
-	}
-};
-
-export const Settings = ( {
-	quantitySelectorStyle,
-	setAttributes,
-}: AddToCartFormSettings ) => {
-	return (
-		<InspectorControls>
-			<PanelBody title={ __( 'Quantity Selector', 'woocommerce' ) }>
-				<ToggleGroupControl
-					className="wc-block-editor-quantity-selector-style"
-					__nextHasNoMarginBottom
-					value={ quantitySelectorStyle }
-					isBlock
-					onChange={ ( value: QuantitySelectorStyle ) => {
-						setAttributes( {
-							quantitySelectorStyle:
-								value as QuantitySelectorStyle,
-						} );
-					} }
-					help={ getHelpText( quantitySelectorStyle ) }
-				>
-					<ToggleGroupControlOption
-						label={ __( 'Input', 'woocommerce' ) }
-						value={ QuantitySelectorStyle.Input }
-					/>
-					<ToggleGroupControlOption
-						label={ __( 'Stepper', 'woocommerce' ) }
-						value={ QuantitySelectorStyle.Stepper }
-					/>
-				</ToggleGroupControl>
-			</PanelBody>
-		</InspectorControls>
-	);
-};
