--- plugins/woocommerce/changelog/47664-42299-improve-styling-for-quantity-selector-and-variation-single-product-template
@@ -1,4 +1,0 @@
-Significance: minor
-Type: add
-
-Add to Cart with Options: Add Quantity Selector option.
