+++ plugins/woocommerce/changelog/48624-revert-add-to-fart-form-stepper
@@ -1,0 +1,4 @@
+Significance: patch
+Type: tweak
+Comment: This PR is just a revert.
+
