+++ plugins/woocommerce/src/Blocks/BlockTypesController.php
@@ -356,6 +356,7 @@ final class BlockTypesController {
 			$block_types = array_diff(
 				$block_types,
 				array(
+					'AddToCartForm',
 					'Breadcrumbs',
 					'CatalogSorting',
 					'ClassicTemplate',
